Thêm chức năng quản lý sản phẩm trong giỏ hàng
Đã thêm controller, service, route và validator cho sản phẩm trong giỏ hàng của khách hàng. Bổ sung endpoint POST để thêm sản phẩm vào giỏ hàng, cập nhật luồng tạo user để tự động tạo cart, và chỉnh sửa logic xóa user. Điều chỉnh import và sử dụng đúng service trong user.controller.

+++ server/src/services/management/user.service.js
@@ -29,6 +29,12 @@ const userService = {
             },
         });
 
+        await prisma.carts.create({
+            data: {
+                user_id: newUser.id
+            }
+        })
+
         return newUser;
     },
 
@@ -92,9 +98,10 @@ const userService = {
         return listUsers;
     },
 
-    deleteUser: async (userId) => {
-
-        await deleteImage(userId, "users", "avatar");
+    deleteUser: async (userId, currentUser) => {
+        if (currentUser.avatar) {
+            await deleteImage(userId, "users", "avatar");
+        }
         await prisma.users.delete({
             where: { id: userId }
         })
